requirements.txt
requests>=2.31.0
cryptography>=41.0.0
websockets>=12.0
python-dotenv>=1.0.0
colorama>=0.4.6
pandas>=2.0.0
